fixup! feat(backup-agent): move baremetal logics in backup-agent module
