+++ packages/manager/modules/backup-agent/package.json
@@ -6,8 +6,59 @@
   "license": "BSD-3-Clause",
   "author": "OVH SAS",
   "sideEffects": false,
-  "main": "dist/src/lib.js",
-  "types": "dist/types/src/lib.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/src/lib",
+      "require": "./dist/src/lib.js",
+      "types": "./dist/types/src/lib.d.ts"
+    },
+    "./package.json": {
+      "import": "./package.json",
+      "require": "./package.json"
+    },
+    "./data/hooks/baremetal/useBaremetalsList": {
+      "import": "./dist/src/data/hooks/baremetal/useBaremetalsList",
+      "require": "./dist/src/data/hooks/baremetal/useBaremetalsList.js",
+      "types": "./dist/types/src/data/hooks/baremetal/useBaremetalsList.d.ts"
+    },
+    "./mocks/baremetals/baremetals.mocks": {
+      "import": "./dist/src/mocks/baremetals/baremetals.mocks",
+      "require": "./dist/src/mocks/baremetals/baremetals.mocks.js",
+      "types": "./dist/types/src/mocks/baremetals/baremetals.mocks.d.ts"
+    },
+    "./mocks/baremetals/baremetals.handler": {
+      "import": "./dist/src/mocks/baremetals/baremetals.handler",
+      "require": "./dist/src/mocks/baremetals/baremetals.handler.js",
+      "types": "./dist/types/src/mocks/baremetals/baremetals.handler.d.ts"
+    },
+    "./components/Fields/RhfField/RhfField.component": {
+      "import": "./dist/src/components/Fields/RhfField.component",
+      "require": "./dist/src/components/Fields/RhfField.component.js",
+      "types": "./dist/types/src/components/Fields/RhfField.component.d.ts"
+    },
+    "./components/Fields/RhfInput": {
+      "import": "./dist/src/components/Fields/RhfInput",
+      "require": "./dist/src/components/Fields/RhfInput.js",
+      "types": "./dist/types/src/components/Fields/RhfInput.d.ts"
+    },
+    "./components/Fields/RhfQuantity": {
+      "import": "./dist/src/components/Fields/RhfQuantity",
+      "require": "./dist/src/components/Fields/RhfQuantity.js",
+      "types": "./dist/types/src/components/Fields/RhfQuantity.d.ts"
+    },
+    "./components/Fields/RhfCombobox": {
+      "import": "./dist/src/components/Fields/RhfCombobox",
+      "require": "./dist/src/components/Fields/RhfCombobox.js",
+      "types": "./dist/types/src/components/Fields/RhfCombobox.d.ts"
+    },
+    "./components/CommonFields/BaremetalOption/BaremetalOption.component": {
+      "import": "./dist/src/components/CommonFields/BaremetalOption/BaremetalOption.component",
+      "require": "./dist/src/components/CommonFields/BaremetalOption/BaremetalOption.component.js",
+      "types": "./dist/types/src/components/CommonFields/BaremetalOption/BaremetalOption.component.d.ts"
+    }
+  },
+  "main": "./dist/src/lib.js",
+  "types": "./dist/types/src/lib.d.ts",
   "scripts": {
     "build": "tsc && tsc-alias",
     "dev": "tsc && tsc-alias",
