--- packages/manager/modules/backup-agent/src/components/CommonFields/BaremetalOption/index.ts
@@ -1,3 +1,0 @@
-import BaremetalOption from './BaremetalOption.component';
-
-export { BaremetalOption };
