--- packages/manager/modules/backup-agent/src/components/CommonFields/index.ts
@@ -1,1 +1,0 @@
-export * from './BaremetalOption';
