--- packages/manager/modules/backup-agent/src/components/index.ts
@@ -1,2 +1,0 @@
-export * from './CommonFields';
-export * from './Fields';
