--- packages/manager/modules/backup-agent/src/data/api/index.ts
@@ -1,6 +1,0 @@
-export * from './agents/agents.requests';
-export * from './baremetal/baremetals';
-export * from './vaults/vault.requests';
-export * from './tenants/tenants.requests';
-export * from './tenants/backupPolicies.requests';
-export * from './locations/location.requests';
