+++ packages/manager/modules/backup-agent/src/data/api/tenants/backupPolicies.requests.ts
@@ -1,4 +1,4 @@
-import { getDetailsVspcTenantRoute } from '@/data';
+import { getDetailsVspcTenantRoute } from '@/data/api/tenants/tenants.requests';
 
 export const getBackupPoliciesRoute = (tenantId: string) =>
   `${getDetailsVspcTenantRoute(tenantId)}/backupPolicies`;
