--- packages/manager/modules/backup-agent/src/data/hooks/agents/index.ts
@@ -1,1 +1,0 @@
-export * from './getAgents';
