--- packages/manager/modules/backup-agent/src/data/hooks/baremetal/index.ts
@@ -1,1 +1,0 @@
-export * from './useBaremetalsList';
