--- packages/manager/modules/backup-agent/src/data/hooks/index.ts
@@ -1,5 +1,0 @@
-export * from './agents';
-export * from './baremetal';
-export * from './location';
-export * from './tenants';
-export * from './vaults';
