--- packages/manager/modules/backup-agent/src/data/hooks/location/index.ts
@@ -1,1 +1,0 @@
-export * from './getLocationDetails';
