--- packages/manager/modules/backup-agent/src/data/hooks/tenants/index.ts
@@ -1,5 +1,0 @@
-export * from './useBackupTenantDetails';
-export * from './useBackupTenants';
-export * from './useDeleteTenant';
-export * from './useVspcTenants';
-export * from './useVspcTenantBackupPolicies';
