+++ packages/manager/modules/backup-agent/src/data/hooks/tenants/useVspcTenants.ts
@@ -1,5 +1,3 @@
-import { useMemo } from 'react';
-
 import {
   DefinedInitialDataOptions,
   UseQueryResult,
