--- packages/manager/modules/backup-agent/src/data/hooks/vaults/index.ts
@@ -1,3 +1,0 @@
-export * from './getVault';
-export * from './getVaultDetails';
-export * from './useDeleteVault';
