--- packages/manager/modules/backup-agent/src/data/index.ts
@@ -1,2 +1,0 @@
-export * from './api';
-export * from './hooks';
