--- packages/manager/modules/backup-agent/src/mocks/agents/index.ts
@@ -1,2 +1,0 @@
-export * from './agents';
-export * from './agents.handler';
