--- packages/manager/modules/backup-agent/src/mocks/baremetals/index.ts
@@ -1,2 +1,0 @@
-export * from './baremetals.mocks';
-export * from './baremetals.handler';
