--- packages/manager/modules/backup-agent/src/mocks/index.ts
@@ -1,5 +1,0 @@
-export * from './agents';
-export * from './vaults';
-export * from './tenant';
-export * from './location';
-export * from './baremetals';
