--- packages/manager/modules/backup-agent/src/mocks/location/index.ts
@@ -1,2 +1,0 @@
-export * from './locations';
-export * from './locations.handler';
