--- packages/manager/modules/backup-agent/src/mocks/tenant/index.ts
@@ -1,4 +1,0 @@
-export * from './tenants.mock';
-export * from './tenants.handler';
-export * from './vspcTenants.handler';
-export * from './vspcTenants.mock';
