--- packages/manager/modules/backup-agent/src/mocks/vaults/index.ts
@@ -1,2 +1,0 @@
-export * from './vaults';
-export * from './vaults.handler';
