+++ packages/manager/modules/backup-agent/src/pages/services/dashboard/general-information/_components/__tests__/SubscriptionTile.test.tsx
@@ -1,5 +1,4 @@
 import { screen, waitFor } from '@testing-library/react';
-import { vi } from 'vitest';
 
 import { urls } from '@/routes/Routes.constants';
 import { renderTest } from '@/test-utils/Test.utils';
