--- packages/manager/modules/backup-agent/src/types/index.ts
@@ -1,11 +1,0 @@
-export * from './Agent.type';
-export * from './Baremetal.type';
-export * from './Bucket.type';
-export * from './ClientApi.type';
-export * from './Dashboard.type';
-export * from './Location.type';
-export * from './Resource.type';
-export * from './Tenant.type';
-export * from './Utils.type';
-export * from './Vault.type';
-export * from './VspcTenant.type';
